requirements.txt
astroid==1.4.9
coreapi==2.2.4
coreschema==0.0.4
Django==1.10.5
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-filter==1.0.1
django-guardian==1.4.6
djangorestframework==3.5.4
drfdocs==0.0.11
isort==4.2.5
itypes==1.1.0
Jinja2==2.9.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
MarkupSafe==0.23
mccabe==0.6.1
openapi-codec==1.3.1
psycopg2==2.6.2
pylint==1.6.5
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
wrapt==1.10.8
